fix semantic comparison for service
