+++ pkg/reconciler/component/reconciler/render/crd.go
@@ -1,0 +1,21 @@
+package render
+
+import (
+	apiextensionsv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
+	"k8s.io/apimachinery/pkg/version"
+)
+
+func CRDPriotizedVersion(crd *apiextensionsv1.CustomResourceDefinition) *apiextensionsv1.CustomResourceDefinitionVersion {
+	var crdv *apiextensionsv1.CustomResourceDefinitionVersion
+	for _, v := range crd.Spec.Versions {
+		if crdv == nil {
+			crdv = &v
+			continue
+		}
+
+		if version.CompareKubeAwareVersionStrings(v.Name, crdv.Name) > 0 {
+			crdv = &v
+		}
+	}
+	return crdv
+}
